requirements.txt
requests
re
os
json
pyyaml